refactored VC back into objects for herd access + few gui merge bits
